WebChat Java WebSockets
This is a simple web chat that uses a javascript client and a Java
websockets server.  The user connects to the chat, chooses a chat room
and then he can start to chat. He will see only the messagess that are
posted after he enters the chat room. The javascript sends and receives
information about the room (where to send and receive messages)
,message, date in a JSON format via the websocket. The server has a
server endpoint "/chat/{room}", oppens a session and adds the room where
to chat in the "User Properties" for the new user.  It has also a class
to decode the json message from the javascript clients, and store the
message properties and a encode class that will transoform the text into
a json message to be sent. Finaly, the server will pass through all
sessions and send the chat messages to all the clients connected.